requirements.txt
einops==0.8.1
faiss_gpu==1.7.2
h5py==3.12.1
linear_attention_transformer==0.19.1
matplotlib==3.10.0
mne==1.9.0
numpy==2.2.3
pandas==1.5.3
pyhealth==1.1.6
pytorch_lightning==2.4.0
scikit_learn==1.6.1
scipy==1.15.2
torch==2.5.1+cu121
tqdm==4.67.1
